NVA-2411: Refactor package name.
