+++ src/main/java/com/github/bibsysdev/nva/cristin/person/NvaPersonBuilder.java
@@ -1,12 +1,12 @@
-package com.github.bibsysdev.nva_cristin_person;
+package com.github.bibsysdev.nva.cristin.person;
 
-import static com.github.bibsysdev.nva_cristin_person.CristinApiClient.BASE_URL;
-import static com.github.bibsysdev.nva_cristin_person.UriUtils.buildUri;
-import com.github.bibsysdev.nva_cristin_person.model.cristin.CristinAffiliation;
-import com.github.bibsysdev.nva_cristin_person.model.cristin.CristinPerson;
-import com.github.bibsysdev.nva_cristin_person.model.nva.NvaAffiliation;
-import com.github.bibsysdev.nva_cristin_person.model.nva.NvaIdentifier;
-import com.github.bibsysdev.nva_cristin_person.model.nva.NvaPerson;
+import static com.github.bibsysdev.nva.cristin.person.CristinApiClient.BASE_URL;
+import static com.github.bibsysdev.nva.cristin.person.UriUtils.buildUri;
+import com.github.bibsysdev.nva.cristin.person.model.cristin.CristinAffiliation;
+import com.github.bibsysdev.nva.cristin.person.model.cristin.CristinPerson;
+import com.github.bibsysdev.nva.cristin.person.model.nva.NvaAffiliation;
+import com.github.bibsysdev.nva.cristin.person.model.nva.NvaIdentifier;
+import com.github.bibsysdev.nva.cristin.person.model.nva.NvaPerson;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
@@ -55,15 +55,18 @@ public class NvaPersonBuilder {
     private static List<NvaAffiliation> createNvaAffiliationList(CristinPerson cristinPerson) {
         List<NvaAffiliation> nvaAffiliations = new ArrayList<>();
         if (cristinPerson.getAffiliations() != null) {
-            cristinPerson.getAffiliations().stream().filter(CristinAffiliation::getActive).forEach(cristinAffiliation -> {
-                NvaAffiliation nvaAffiliation = new NvaAffiliation();
-                nvaAffiliation.setId(cristinAffiliation.getUnit().getUrl());
-                if (cristinAffiliation.getPosition() != null && !cristinAffiliation.getPosition().isEmpty()) {
-                    nvaAffiliation.setRole(cristinAffiliation.getPosition());
-                }
+            cristinPerson.getAffiliations()
+                .stream()
+                .filter(CristinAffiliation::getActive)
+                .forEach(cristinAffiliation -> {
+                    NvaAffiliation nvaAffiliation = new NvaAffiliation();
+                    nvaAffiliation.setId(cristinAffiliation.getUnit().getUrl());
+                    if (cristinAffiliation.getPosition() != null && !cristinAffiliation.getPosition().isEmpty()) {
+                        nvaAffiliation.setRole(cristinAffiliation.getPosition());
+                    }
 
-                nvaAffiliations.add(nvaAffiliation);
-            });
+                    nvaAffiliations.add(nvaAffiliation);
+                });
         }
         return nvaAffiliations;
     }
