+++ src/main/java/com/github/bibsysdev/nva/cristin/person/model/nva/EmptyNvaPerson.java
@@ -1,4 +1,4 @@
-package com.github.bibsysdev.nva_cristin_person.model.nva;
+package com.github.bibsysdev.nva.cristin.person.model.nva;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,6 +7,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 public class EmptyNvaPerson extends NvaPerson {
     /*
     TODO: NP-2315: Here we can return custom fields like "status": 404 to show that lookup of
-     project with given ID did not match any results or that project contained invalid/insufficient data
+     person with given ID did not match any results or that person contained invalid/insufficient data
     */
 }
