+++ src/main/java/com/github/bibsysdev/nva/cristin/person/model/nva/NvaPerson.java
@@ -1,8 +1,8 @@
-package com.github.bibsysdev.nva_cristin_person.model.nva;
+package com.github.bibsysdev.nva.cristin.person.model.nva;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS;
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
-import static com.github.bibsysdev.nva_cristin_person.PersonsWrapper.CONTEXT;
+import static com.github.bibsysdev.nva.cristin.person.PersonsWrapper.CONTEXT;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
